+++ applications/fishing-map/src/features/map/map.hooks.ts
@@ -24,8 +24,10 @@ import { useLocationConnect } from 'routes/routes.hook'
 import { DEFAULT_WORKSPACE_ID, WorkspaceCategories } from 'data/workspaces'
 import useMapInstance from 'features/map/map-context.hooks'
 import { removeDatasetVersion } from 'features/datasets/datasets.utils'
+import { PRESENCE_POC_INTERACTION, USE_PRESENCE_POC } from 'features/datasets/datasets.slice'
 import { useTimerangeConnect } from 'features/timebar/timebar.hooks'
 import { selectHighlightedEvent, setHighlightedEvent } from 'features/timebar/timebar.slice'
+import { isGuestUser } from 'features/user/user.selectors'
 import { t } from 'features/i18n/i18n'
 import {
   selectDefaultMapGeneratorsConfig,
@@ -78,6 +80,7 @@ export const useGeneratorsConnect = () => {
 export const useClickedEventConnect = () => {
   const map = useMapInstance()
   const dispatch = useDispatch()
+  const guestUser = useSelector(isGuestUser)
   const clickedEvent = useSelector(selectClickedEvent)
   const locationType = useSelector(selectLocationType)
   const fishingInteractionStatus = useSelector(selectFishingInteractionStatus)
@@ -172,15 +175,15 @@ export const useClickedEventConnect = () => {
     // get temporal grid clicked features and order them by sublayerindex
     const fishingActivityFeatures = event.features
       .filter((feature) => {
-        if (!feature.temporalgrid) {
+        if (!feature.temporalgrid?.visible) {
           return false
         }
-        return (
-          feature.temporalgrid.visible &&
-          SUBLAYER_INTERACTION_TYPES_WITH_VESSEL_INTERACTION.includes(
-            feature.temporalgrid.sublayerInteractionType
-          )
+        const hasSubLayerInteraction = SUBLAYER_INTERACTION_TYPES_WITH_VESSEL_INTERACTION.includes(
+          feature.temporalgrid.sublayerInteractionType
         )
+        const isPresencePOCFeature =
+          feature.temporalgrid.sublayerInteractionType === PRESENCE_POC_INTERACTION
+        return hasSubLayerInteraction || (USE_PRESENCE_POC && isPresencePOCFeature && !guestUser)
       })
       .sort((feature) => feature.temporalgrid?.sublayerIndex ?? 0)