+++ applications/vessel-history/src/features/datasets/datasets.selectors.ts
@@ -2,6 +2,7 @@ import { createSelector } from 'reselect'
 import { uniqBy } from 'lodash'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 import { Dataset, DatasetTypes } from '@globalfishingwatch/api-types'
+import { FULL_SUFIX } from 'data/config'
 import { selectDatasets } from './datasets.slice'
 
 export const getDatasetsByDataview = (dataview: UrlDataviewInstance) =>
@@ -44,6 +45,23 @@ export const selectTracksDatasets = createSelector(
   }
 )
 
-export const getRelatedDatasetByType = (dataset?: Dataset, datasetType?: DatasetTypes) => {
+export const getRelatedDatasetByType = (
+  dataset?: Dataset,
+  datasetType?: DatasetTypes,
+  fullDatasetAllowed = false
+) => {
+  if (fullDatasetAllowed) {
+    const fullDataset = dataset?.relatedDatasets?.find(
+      (relatedDataset) =>
+        relatedDataset.type === datasetType && relatedDataset.id.startsWith(FULL_SUFIX)
+    )
+    if (fullDataset) {
+      return fullDataset
+    }
+  }
   return dataset?.relatedDatasets?.find((relatedDataset) => relatedDataset.type === datasetType)
+}
+
+export const getRelatedDatasetsByType = (dataset?: Dataset, datasetType?: DatasetTypes) => {
+  return dataset?.relatedDatasets?.filter((relatedDataset) => relatedDataset.type === datasetType)
 }