+++ applications/vessel-history/src/features/vessel-list-item/VesselListItem.tsx
@@ -26,7 +26,10 @@ const VesselListItem: React.FC<ListItemProps> = (props): React.ReactElement => {
   const sourceAPI = getVesselAPISource(vessel)
 
   return (
-    <div className={styles.vesselItem}>
+    <Link
+      to={['profile', vessel.dataset ?? 'NA', vessel.id ?? 'NA', vessel.vesselMatchId ?? 'NA']}
+      className={styles.vesselItem}
+    >
       {props.saved && onDeleteClick && (
         <IconButton
           type="warning"
@@ -36,11 +39,7 @@ const VesselListItem: React.FC<ListItemProps> = (props): React.ReactElement => {
           onClick={onDeleteClick}
         ></IconButton>
       )}
-      <Link
-        to={['profile', vessel.dataset ?? 'NA', vessel.id ?? 'NA', vessel.vesselMatchId ?? 'NA']}
-      >
-        <h3>{vessel?.shipname}</h3>
-      </Link>
+      <h3>{vessel?.shipname}</h3>
       <div className={styles.identifiers}>
         <div>
           <label>{t('vessel.flag', 'flag')}</label>
@@ -98,7 +97,7 @@ const VesselListItem: React.FC<ListItemProps> = (props): React.ReactElement => {
           </div>
         )}
       </div>
-    </div>
+    </Link>
   )
 }