+++ packages/dataviews-client/src/index.ts
@@ -13,3 +13,4 @@ export {
   MERGED_ACTIVITY_ANIMATED_HEATMAP_GENERATOR_ID,
 } from './resolve-dataviews-generators'
 export { default as resolveEndpoint } from './resolve-endpoint'
+export * from './resources'