+++ packages/dataviews-client/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @globalfishingwatch/dataviews-client
 
+## 11.10.1
+
+### Patch Changes
+
+- fix release
+
+## 11.10.0
+
+### Minor Changes
+
+- cb0c6c42: resources slice and selectors
+
+## 11.9.0
+
+### Minor Changes
+
+- 0c82c8d3: support presence-POC interactionType
+-
+
+## 11.8.1
+
+### Patch Changes
+
+- Include resources slice, selectors and helpers to consume API
+
 ## 11.8.0
 
 ### Minor Changes